# openapi/aggregation-api.yaml
openapi: 3.0.1
info:
  title: Aggregation Service API
  description: Service provided aggregation data from multiple databases.
  version: 1.0.0
tags:
  - name: users
paths:
  /v1/users:
    get:
      description: Retrieve all available users.
      tags:
        - users
      operationId: receiveUsers
      parameters:
        - $ref: '#/components/parameters/id'
        - $ref: '#/components/parameters/username'
        - $ref: '#/components/parameters/name'
        - $ref: '#/components/parameters/surname'
      responses:
        '200':
          description: List of users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponseDto'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /v2/users:
    get:
      description: Retrieve all available users.
      tags:
        - users
      operationId: receiveUsers2
      parameters:
        - $ref: '#/components/parameters/id'
        - $ref: '#/components/parameters/username'
        - $ref: '#/components/parameters/name'
        - $ref: '#/components/parameters/surname'
      responses:
        '200':
          description: List of users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponseDto'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
components:
  parameters:
    id:
      name: id
      in: query
      schema:
        type: string
      description: 'Unique user identifier per tenant.'
    username:
      name: username
      in: query
      schema:
        type: string
      description: 'User username.'
    name:
      name: name
      in: query
      schema:
        type: string
      description: 'User name.'
    surname:
      name: surname
      in: query
      schema:
        type: string
      description: 'User surname.'
  schemas:
    UsersParameters:
      type: object

    UserResponseDto:
      type: object
      properties:
        id:
          type: string
          description: "System user identifier."
          example: "12345qwerty"
        username:
          type: string
          description: "System user name."
          example: "test_user"
        name:
          type: string
          description: "User name."
          example: "Jack"
        surname:
          type: string
          description: "User surname."
          example: "Sparrow"
    ErrorResponseDto:
      type: object
      properties:
        errorCode:
          type: string
        errorMessage:
          type: string